requirements.txt
line-bot-sdk==1.3.0
gunicorn==19.7.1
Flask==0.12.2
